% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicts.R
\name{hist.semicts}
\alias{hist.semicts}
\title{Generates a histogram for the semi-continuous data provided}
\usage{
\method{hist}{semicts}(obj, xlab = "X", ylab = "Density", main = "",
  cols = c(rgb(0, 0, 1, 0.5), rgb(1, 0, 0, 0.5)),
  legends = c("Intensity", "Proportion of Zeros"))
}
\arguments{
\item{obj}{an array of semi-continuous data}

\item{xlab}{label for x-axis (has a default)}

\item{ylab}{label for y-axis (has a default)}

\item{main}{title (has a default)}

\item{cols}{two colors as an array: one for the proportion of zeroes, and the other for the continuous data (has a default)}

\item{legends}{names for the legend as an array of two elements (has a default)}
}
\value{
histogram object
}
\description{
Generates a histogram for the semi-continuous data provided
}
\examples{
x <- rsemicts(100, pzero=0.4, cts.density="lnorm", cts.params=c(1,1))
hist(x)
}
