% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicts.R
\name{pred.perf}
\alias{pred.perf}
\title{Returns MPSE, MAD, average of positive predictions when the true value is zero
(and vice-versa), proportion of matched zeroes, and proportion of matched positives.}
\usage{
pred.perf(y_pred, y)
}
\arguments{
\item{y_pred}{A semicts object (for ex. returned from the rsemicts function)}

\item{y}{A semicts object}
}
\description{
Returns MPSE, MAD, average of positive predictions when the true value is zero
(and vice-versa), proportion of matched zeroes, and proportion of matched positives.
}
\examples{
x <- rsemicts(100, pzero=0.4, cts.density="lnorm", cts.param=c(1,1))
y <- rsemicts(100, pzero=0.4, cts.density="lnorm", cts.param=c(1,1))
pred.perf(x, y)
}
