% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semicts.R
\name{rtobit}
\alias{rtobit}
\title{Generate a random sample from a latent normal distribution (tobit)}
\usage{
rtobit(n, mean = 1, sd = 1)
}
\arguments{
\item{n}{Number of random variables to generate}

\item{mean}{mean of the latent normal distribution}

\item{sd}{standard deviation of the latent normal distribution}
}
\value{
An array of semi-continuous random variables.
}
\description{
Generate a random sample from a latent normal distribution (tobit)
}
\examples{
rtobit(100)
}

